Refactoring models
