+++ packages/main-api/src/models/dynamo_tables/main/email/mod.rs
@@ -1,3 +1,5 @@
 pub mod email_verification;
+#[cfg(test)]
+mod tests;
 
 pub use email_verification::*;
