+++ packages/main-api/src/models/dynamo_tables/main/user/mod.rs
@@ -1,3 +1,18 @@
 pub mod user;
+pub mod user_evm_address;
+pub mod user_metadata;
+pub mod user_phone_number;
+pub mod user_principal;
+pub mod user_referral_code;
+pub mod user_telegram;
 
 pub use user::*;
+pub use user_evm_address::*;
+pub use user_metadata::*;
+pub use user_phone_number::*;
+pub use user_principal::*;
+pub use user_referral_code::*;
+pub use user_telegram::*;
+
+#[cfg(test)]
+mod tests;
