+++ packages/main-api/src/models/dynamo_tables/main/user/user_metadata.rs
@@ -1,0 +1,13 @@
+use super::*;
+use bdk::prelude::*;
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, DynamoEntity)]
+#[serde(untagged)]
+pub enum UserMetadata {
+    User(User),
+    UserPrincipal(UserPrincipal),
+    UserEvmAddress(UserEvmAddress),
+    UserReferralCode(UserReferralCode),
+    UserPhoneNumber(UserPhoneNumber),
+    UserTelegram(UserTelegram),
+}
