+++ packages/main-api/src/models/dynamo_tables/main/user/user_principal.rs
@@ -1,0 +1,20 @@
+use crate::types::*;
+use bdk::prelude::*;
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, DynamoEntity)]
+pub struct UserPrincipal {
+    pub pk: Partition,
+    #[dynamo(index = "gsi1", sk)]
+    pub sk: EntityType,
+
+    #[dynamo(name = "find_by_principal", prefix = "PRINCIPAL", index = "gsi1", pk)]
+    pub principal: String,
+}
+
+impl UserPrincipal {
+    pub fn new(pk: Partition, principal: String) -> Self {
+        let sk = EntityType::UserPrincipal;
+
+        Self { pk, sk, principal }
+    }
+}
